# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api SIAC",
        "description": "SIAC",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Crear nuevo usuario",
                "description": "Crear nuevo usuario",
                "operationId": "d764dd091cc4494ae0baf360b03319f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "ejemplo@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Record created successfully"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Iniciar Sesión",
                "description": "Iniciar Sesión",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sesión iniciada correctamente"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Cerrar Sesión",
                "description": "Cerrar Sesión",
                "operationId": "a8dd056be3e98a38c657f73ed6cdef2d",
                "requestBody": {
                    "required": false
                },
                "responses": {
                    "200": {
                        "description": "Sesión iniciada correctamente"
                    }
                }
            }
        }
    }
}